backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
jsonschema==4.20.0